style: Reflect feedback
