+++ front/package.json
@@ -4,8 +4,8 @@
 	"private": true,
 	"dependencies": {
 		"@testing-library/jest-dom": "^5.11.4",
-		"@testing-library/react": "^11.1.0",
-		"@testing-library/user-event": "^12.1.10",
+		"@testing-library/react": "^12.0.0",
+		"@testing-library/user-event": "^13.2.1",
 		"antd": "^4.16.11",
 		"react": "^17.0.2",
 		"react-dom": "^17.0.2",
@@ -14,7 +14,7 @@
 		"react-slick": "^0.28.1",
 		"slick-carousel": "^1.8.1",
 		"styled-components": "^5.3.0",
-		"web-vitals": "^1.0.1"
+		"web-vitals": "^2.1.0"
 	},
 	"scripts": {
 		"start": "react-scripts start",
