+++ front/src/app.jsx
@@ -1,6 +1,7 @@
 import React from "react";
 import { Switch, Route } from "react-router-dom";
 import Home from "../src/pages/Home";
+import "antd/dist/antd.css";
 
 const App = () => {
 	return (
