+++ front/src/components/home/FeedList.jsx
@@ -1,6 +1,5 @@
 import React from "react";
 import { Row, Col } from "antd";
-import "antd/dist/antd.css";
 import FeedCard from "./FeedCard";
 
 const FeedList = () => {
